# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: BarkBase API
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  time:
                    type: string
                    format: date-time
  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login success
  /api/v1/bookings:
    get:
      summary: List bookings
      parameters:
        - in: header
          name: x-tenant-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of bookings
    post:
      summary: Create booking
      parameters:
        - in: header
          name: x-tenant-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        "201":
          description: Booking created
  /api/v1/pets:
    get:
      summary: List pets
      parameters:
        - in: header
          name: x-tenant-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of pets
components:
  schemas:
    BookingCreate:
      type: object
      required:
        - petId
        - ownerId
        - checkIn
        - checkOut
        - segments
      properties:
        petId:
          type: string
        ownerId:
          type: string
        status:
          type: string
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        depositCents:
          type: integer
        segments:
          type: array
          items:
            type: object
            required:
              - kennelId
              - startDate
              - endDate
            properties:
              kennelId:
                type: string
              startDate:
                type: string
                format: date-time
              endDate:
                type: string
                format: date-time